Order Management B2B contact sla min count added

+++ src/main/java/com/xyz/bd/webmaster/modules/orders/OrderModelEntity.java
@@ -110,7 +110,7 @@ public class OrderModelEntity extends BaseEntity implements Serializable {
 
     @Column(name = "contact_sla_min")
     @JsonProperty("contactSlaMin")
-    private Integer contactSlaMin;
+    private Long contactSlaMin;
 
     @Column(name = "installation_sla_min")
     @JsonProperty("installationSlaMin")
